# Optimize sequence conversions by avoiding intermediate PyValue allocations