--- a/bench/scripts/run.py
+++ b/bench/scripts/run.py
@@ -46,16 +46,16 @@ def compile(bench, compilecmd):
 
 sbt = where('sbt')
 
-sizes = [1, 5, 10, 50, 100]
+sizes = [2 ** x for x in range(1, 15)]
 
 benchmarks = [
     ('CreatePythonCopyBenchmark', sizes, None),
-    ('CreatePythonProxyBenchmark', sizes, None),
-    ('SumPythonCopyBenchmark', sizes, None),
-    ('SumPythonProxyBenchmark', sizes, None),
-    ('SumScalaBenchmark', sizes, None),
-    ('TensorFlowAppScalaPyBenchmark', [0], 50),
-    ('TensorFlowAppPythonBenchmark', [0], 50),
+    # ('CreatePythonProxyBenchmark', sizes, None),
+    # ('SumPythonCopyBenchmark', sizes, None),
+    # ('SumPythonProxyBenchmark', sizes, None),
+    # ('SumScalaBenchmark', sizes, None),
+    # ('TensorFlowAppScalaPyBenchmark', [0], 50),
+    # ('TensorFlowAppPythonBenchmark', [0], 50),
 ]
 
 configurations = [