--- a/core/native/src/main/scala/me/shadaj/scalapy/interpreter/ThreadLocal.scala
+++ b/core/native/src/main/scala/me/shadaj/scalapy/interpreter/ThreadLocal.scala
@@ -1,0 +1,9 @@
+package me.shadaj.scalapy.interpreter
+
+private[scalapy] class SingleThreadLocal[T](value: T) {
+  def get(): T = value
+}
+
+object SingleThreadLocal {
+  def withInitial[T](initial: () => T) = new SingleThreadLocal[T](initial())
+}