--- a/core/shared/src/main/scala/me/shadaj/scalapy/py/Any.scala
+++ b/core/shared/src/main/scala/me/shadaj/scalapy/py/Any.scala
@@ -22,7 +22,6 @@ trait Any { self =>
 
   final def as[T: Reader]: T = implicitly[Reader[T]].read(value)
 
-
   final def del(): Unit = {
     value.cleanup()
     cleaned = true