before:

name | size | jvm | scala-native | bytes | gb/s (jvm) | gb/s (scala native) | mb/s (jvm) | mb/s (scala native)
-- | -- | -- | -- | -- | -- | -- | -- | --
CreatePythonCopyBenchmark | 2 | 17643 | 3917 | 16 | 0.000906875248 | 0.004084758744 | 0.906875248 | 4.084758744
CreatePythonCopyBenchmark | 4 | 19567 | 4408 | 32 | 0.001635406552 | 0.007259528131 | 1.635406552 | 7.259528131
CreatePythonCopyBenchmark | 8 | 28183 | 4388 | 64 | 0.002270872512 | 0.01458523245 | 2.270872512 | 14.58523245
CreatePythonCopyBenchmark | 16 | 37590 | 7174 | 128 | 0.003405160947 | 0.01784220797 | 3.405160947 | 17.84220797
CreatePythonCopyBenchmark | 32 | 61414 | 10920 | 256 | 0.004168430651 | 0.02344322344 | 4.168430651 | 23.44322344
CreatePythonCopyBenchmark | 64 | 185467 | 14617 | 512 | 0.002760598921 | 0.03502770746 | 2.760598921 | 35.02770746
CreatePythonCopyBenchmark | 128 | 325839.5 | 25498 | 1024 | 0.00314265152 | 0.04016001255 | 3.14265152 | 40.16001255
CreatePythonCopyBenchmark | 256 | 653232 | 49453 | 2048 | 0.00313518015 | 0.04141305886 | 3.13518015 | 41.41305886
CreatePythonCopyBenchmark | 512 | 1336270 | 97473 | 4096 | 0.00306524879 | 0.04202189324 | 3.06524879 | 42.02189324

all larger sizes OOM-ed on JVM

after:

name | size | jvm | scala-native | bytes | gb/s (jvm) | gb/s (scala native) | mb/s (jvm) | mb/s (scala native)
-- | -- | -- | -- | -- | -- | -- | -- | --
CreatePythonCopyBenchmark | 2 | 16040 | 3096 | 16 | 0.0009975062344 | 0.005167958656 | 0.9975062344 | 5.167958656
CreatePythonCopyBenchmark | 4 | 17032 | 3206 | 32 | 0.001878816346 | 0.00998128509 | 1.878816346 | 9.98128509
CreatePythonCopyBenchmark | 8 | 19226 | 3396 | 64 | 0.003328825549 | 0.01884570082 | 3.328825549 | 18.84570082
CreatePythonCopyBenchmark | 16 | 23354 | 3787 | 128 | 0.00548085981 | 0.03379984156 | 5.48085981 | 33.79984156
CreatePythonCopyBenchmark | 32 | 31448 | 4549 | 256 | 0.008140422284 | 0.05627610464 | 8.140422284 | 56.27610464
CreatePythonCopyBenchmark | 64 | 48501 | 6021 | 512 | 0.01055648337 | 0.08503570835 | 10.55648337 | 85.03570835
CreatePythonCopyBenchmark | 128 | 84919 | 9568 | 1024 | 0.01205854991 | 0.1070234114 | 12.05854991 | 107.0234114
CreatePythonCopyBenchmark | 256 | 153758 | 17102 | 2048 | 0.01331963215 | 0.1197520758 | 13.31963215 | 119.7520758
CreatePythonCopyBenchmark | 512 | 300012 | 31790 | 4096 | 0.01365278722 | 0.1288455489 | 13.65278722 | 128.8455489
CreatePythonCopyBenchmark | 1024 | 589314 | 61324 | 8192 | 0.01390090851 | 0.1335855456 | 13.90090851 | 133.5855456
CreatePythonCopyBenchmark | 2048 | 1027259 | 120525 | 16384 | 0.01594923968 | 0.1359386019 | 15.94923968 | 135.9386019
CreatePythonCopyBenchmark | 4096 | 1979287 | 241682 | 32768 | 0.01655545659 | 0.1355831216 | 16.55545659 | 135.5831216
CreatePythonCopyBenchmark | 8192 | 3914671 | 486761 | 65536 | 0.01674112588 | 0.1346369163 | 16.74112588 | 134.6369163
CreatePythonCopyBenchmark | 16384 | 7881642 | 1020505.5 | 131072 | 0.01663003724 | 0.1284383083 | 16.63003724 | 128.4383083